🐞fix(M  fly.sh)

+++ fly.sh
@@ -361,9 +361,7 @@ _install_zsh() {
         sed -i -e "/^ZSH_THEME/s/robbyrussell/ys/" "$HOME"/.zshrc
         sed -i -e '/^plugins=.*/s//plugins=\(git z extract docker docker-compose\)/' ~/.zshrc
     fi
-    if _command_exists fzf; then
-        :
-    else
+    if [[ ${apt_update:-0} -eq 1 ]] && ! _command_exists fzf; then
         $cmd install -yqq fzf
         sed -i -e "/^plugins=\(git\)/s/git/git z extract fzf docker-compose/" "$HOME"/.zshrc
     fi
